requirements.txt
gradio==4.12.0
langchain==0.0.354
langchain_core==0.1.5
networkx==2.8.4
python-dotenv==1.0.0
openai==1.6.1
